requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-allauth==65.10.0
pillow==11.3.0
django-cors-headers==4.7.0
python-dotenv==1.1.1
requests==2.31.0